+++ Source/Application/Kysect.Shreks.Application.Handlers/Google/GetCoursePointsBySubjectCourseHandler.cs
@@ -2,7 +2,7 @@
 using Kysect.Shreks.Application.Dto.Study;
 using Kysect.Shreks.Application.Dto.Tables;
 using Kysect.Shreks.Application.Dto.Users;
-using Kysect.Shreks.Application.Specifications;
+using Kysect.Shreks.Application.Extensions;
 using Kysect.Shreks.Core.Models;
 using Kysect.Shreks.Core.Study;
 using Kysect.Shreks.Core.Submissions;
